requirements.txt
# Web Framework
streamlit>=1.28.0

# Google AI Studio API
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0

# Image Processing
Pillow==11.3.0
opencv-python==4.12.0.88

# Environment and Configuration
python-dotenv==1.1.1

# Data Processing
numpy==2.2.6

# HTTP Requests
requests==2.32.5

# Security and Authentication
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2

# Development Tools
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.1

# Google API Dependencies
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.30.0
uritemplate==4.2.0
protobuf==5.29.5
proto-plus==1.26.1

# Pydantic and Validation
pydantic==2.11.7
pydantic_core==2.33.2
annotated-types==0.7.0

# Utility Libraries
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
pyparsing==3.2.3

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=1.0.0

# Visualization
matplotlib>=3.7.0
scipy>=1.10.0
